+++ test/integration/server-57/server_spec.rb
@@ -1,3 +1,3 @@
 require_relative '../helpers/server_spec.rb'
 
-server_test('5.7')
+server_test('5.7', 'server')